+++ src/main/java/org/songeun/petdongne_server/survey/domain/entity/SurveyQuestion.java
@@ -1,4 +1,4 @@
-package org.songeun.petdongne_server.survey.domain;
+package org.songeun.petdongne_server.survey.domain.entity;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
