+++ src/main/java/org/songeun/petdongne_server/survey/infrastructure/SurveyOptionRepository.java
@@ -1,6 +1,6 @@
 package org.songeun.petdongne_server.survey.infrastructure;
 
-import org.songeun.petdongne_server.survey.domain.SurveyOption;
+import org.songeun.petdongne_server.survey.domain.entity.SurveyOption;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
