+++ src/main/java/org/songeun/petdongne_server/survey/infrastructure/SurveyResponseRepository.java
@@ -1,6 +1,6 @@
 package org.songeun.petdongne_server.survey.infrastructure;
 
-import org.songeun.petdongne_server.survey.domain.SurveyResponse;
+import org.songeun.petdongne_server.survey.domain.entity.SurveyResponse;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
